requirements.txt
numpy
scipy
scikit-learn
paho-mqtt
hdbscan
SQLAlchemy
python-dotenv